requirements.txt
# Streamlit
streamlit>=1.40.0

# AI/ML
agno>=2.5.0
google-genai>=1.0.0
anthropic>=0.40.0

# HTTP Requests
requests>=2.31.0

# Google APIs (for JIRA, Docs, Sheets integration)
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Data visualization
plotly>=5.17.0

# Database
# SQLite comes with Python, no need to install

# Image processing (for Figma/Design Compliance)
Pillow>=10.0.0

# Document processing (for BRD Pipeline)
PyPDF2>=3.0.0
python-docx>=1.1.0
openpyxl>=3.1.0

# Utilities
python-dateutil>=2.8.2
